requirements.txt
# 核心数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# 加密货币交易
ccxt>=4.0.0
python-binance>=1.0.17

# 技术分析
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# 回测框架
backtrader>=1.9.78.123
backtesting>=0.3.3
vectorbt>=0.25.0

# 机器学习（基础）
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# 深度学习（计划中，可选）
# tensorflow>=2.13.0
# torch>=2.0.0

# 数据获取
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
yfinance>=0.2.28

# 新闻和情绪分析
newsapi-python>=0.2.7
tweepy>=4.14.0
praw>=7.7.0

# 自然语言处理
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.2

# Jupyter环境
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.14.0
ipywidgets>=8.0.0

# 数据库
sqlalchemy>=2.0.0
pymongo>=4.4.0
redis>=4.6.0

# 配置管理
python-dotenv>=1.0.0
pyyaml>=6.0

# 日志和监控
loguru>=0.7.0

# 时间处理
python-dateutil>=2.8.0
pytz>=2023.3

# HTTP客户端
aiohttp>=3.8.0
httpx>=0.24.0

# 数据验证
pydantic>=2.0.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码质量
black>=23.7.0
flake8>=6.0.0
pylint>=2.17.0

# 性能优化
numba>=0.57.0
cython>=3.0.0

# 实用工具
tqdm>=4.65.0
click>=8.1.0
rich>=13.5.0
